Refactor .js files to .ts and extract code
