--- content-scripts/modules/options/addStyles.js
@@ -1,13 +1,0 @@
-import removeElement from "../utilities/removeElement";
-
-// Utility function to inject CSS into page
-export default function addStyles(id, css) {
-    // First remove before adding
-    removeElement(id);
-
-    const head = document.querySelector("head");
-    const style = document.createElement("style");
-    style.id = id;
-    style.textContent = `${css}`;
-    head.appendChild(style);
-}
