+++ content-scripts/modules/options/constructNewData.ts
@@ -1,12 +1,19 @@
 // Utility function to create data for `injectAllChanges()`
-export default function constructNewData(changes) {
+
+type Changes = {
+    [key: string]: chrome.storage.StorageChange;
+};
+
+export default function constructNewData(changes: Changes): {
+    [key: string]: unknown;
+} {
     // Creates an array of objects from changes
     // The value of each object is the new value
-    const newValuesArray = Object.entries(changes).map((item) => {
-        const itemKey = item[0];
-        const itemValue = item[1]?.newValue;
-        return { [itemKey]: itemValue };
-    });
+    const newValuesArray = Object.entries(changes).map(
+        ([itemKey, itemValue]) => {
+            return { [itemKey]: itemValue?.newValue };
+        },
+    );
 
     // Recreate a hash map to pass to `injectAllChanges()`
     const newChangesData = Object.fromEntries(
