+++ content-scripts/modules/options/index.ts
@@ -1,9 +1,15 @@
 import { useNavBar, hideShortcuts, changeFont } from "./options";
 
 // Array of user preferences, passed to `injectAllChanges`
-export const userPreferences = ["navbar", "shortcuts", "autoLogin", "font"];
+export const userPreferences: string[] = [
+    "navbar",
+    "shortcuts",
+    "autoLogin",
+    "font",
+];
 
-export const injectAllChanges = (data) => {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const injectAllChanges = (data: any): void => {
     Promise.all([
         hideShortcuts(data?.shortcuts),
         useNavBar(data?.navbar),
